backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
motor>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
aiofiles>=23.2.0
cloudinary>=1.40.0
authlib>=1.2.1
httpx>=0.25.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
